services/extractor/requirements.txt
pymupdf
pytesseract
pdf2image
pillow
pydantic
python-dotenv
